fix: bug in individual sequence generation

+++ VRP.FGA_2022/SecondIteration/IndividualSecond.h
@@ -11,7 +11,7 @@ class IndividualSecond
 public:
 	Chromosome chromosome; // on the basis of which chromosome the individual is generated
 	std::vector<int32_t> sequences; // sequences of vertex
-	int64_t fitness;
+	double fitness;
 
 	IndividualSecond(); // null individual
 	IndividualSecond(Chromosome& chromosome, InputData& input, Chromosome& blueprint, double globalLR, bool enable_blueprint);
@@ -19,7 +19,7 @@ public:
 	bool operator < (const IndividualSecond& str) const;
 
 private:
-	int64_t CalculateFitness(InputData& input);
+	double CalculateFitness(InputData& input);
 
 	//generate sequences based on a chromosome
 	vector<int32_t> BornAnIndividual(Chromosome blueprint, double globalLR, bool enable_blueprint);
